Also init spelled-out number from transcription
A second constructor is added to the class, accepting a locale and
string representation of a number, parsing it to construct the actual
numeric value.

As for the original constructor, the fields with a corresponding
argument are initialized with it; the third field is initialized using
a default temporary value (in this case: NaN).