Fix MD syntax in README
The code backticks are moved inside the link brackets, not the other way
around.
Github displays the link in backticks as code instead of as a link, and
I believe it is correct in doing so. It seems more sensible. It must be
the Eclipse MD previewer that handles this incorrectly: it interprets
the content of backticks (here, as a link) but does not for the content
of a link label (here, it displays the backticks and their content as
plain text, instead of as code).