Externalize the source of test inputs
The long and growing list of test input values is moved out of the test
class into a test resource file, that is read and parsed as CSV in the
input source method.
The return type of the latter is changed from stream to iterable,
because the values have to be eagerly parsed before the input stream is
closed (and a stream is lazy, and so would fetch the values too late).