Let JUnit take care of the file loading
The test input source is changed from method to CSV file in the
annotation to the test method. This allows to remove entirely the CSV
loading method.
However this change was not without impact: for it to work properly, the
order of the arguments to the parser test method has to be changed to
match the CSV value order.