Make test for minus zero case-specific
AS "minus zero" is a particular IEEE-754 value, it should be processed
specifically:

Unquestionably, parsing the string "zero" should result in an unsigned
zero number; however, "minus zero" is an explicit request and the parser
should comply by returning this special value.
Spelling it should simply result in the text zero, as both values are
defined as equal in the IEEE-754 standard, and "minus zero" has no
meaning in everyday language; moreover, floating-point operations may
result in this value being produced, so the speller receiving this value
should not treat it as an explicit user request.