Rework the refactored internal function
Transcription of the tens is no longer performed through a call to the
refactored function, but with an actual block (a modification of the
refactored function's body).
The refactored function is changed to better fit the transcription of
hundreds and thousands (and possibly higher ranks), with less and better
named arguments.