Handle all numbers representable as long integers
Yet another array is added: it contains the prefixes of the names of
ranks in the numeric "short" scale.
The rank parameter of the refactored function is retyped to long
integer, because with such high ranks, the values will overflow the
range of ints.
These ranks are handled in a loop rather than sequentially as for the
hundreds and thousands because there is a nice systematic construction
of their names (and it will be even better when we reach
"novemvigintillion" and the likes, yay!).
The check for whether the values in in the longs range is fixed: the
strict inequality is changed to lower-or-equal comparison.