Fix the decimal part check
The function used to check whether the given number has a decimal part
is updated: Math.rint returns an integral double possibly bigger than
its argument, and that is wrong for our use case. Math.floor instead
returns an integral double always lower than its argument (or equal to
it, if it is already integral).