Extract a function to spell out numbers < 1000
Similarly to the hundreds function, this function is cheaper than
spellOutAsLong for numbers in this range.

It is an extraction of the trailing block of spellOutAsLong, and is
called instead of this one in the loop iterating over high ranks, and
twice in the block for numbers between one thousand and one million
(once for the thousands rank, and once for the units one).