Rename "default" parser/speller classes
The default locale is not always en-US, it isn't even *usually* en-US
(unless for US users). The classes now explicitly refer to the en-US
locale, in the value they return for getSupportedLocale and their class
name.

The locale used to format the decimal part to transcribe is also
updated.

The test classes are also modified to match.