Invert flow when checking remainder
The control flow of the algorithm is simplified: instead of returning
when the remainder is not empty after performing an operation almost
similar to one performed just below, after the parent if block, the
code now returns when the remainder is zero (meaning that the
transcription is written out entirely), and otherwise just falls
through to the original operation, now unduplicated, after updating a
variable if necessary.

Well anyway. The code now reads much more naturally and overall makes
more sense