Hide parameters from method prototype
Since the method spelling a long is no longer recursive, its rank and
index parameters are removed, replaced in their use with local
variables initialized with the values previously given to the arguments
in the only invocation of the method left.

The variables are also commented to explain their values.