Implement big numbers transcription, in theory
The code is not yet tested, as the most crucial part, getting the rank
names, is not implemented yet and thus tests fail immediately. However,
it is extremely similar to the long-spelling algorithm (apart from using
doubles instead of longs) and therefore I suspect that it will work just
fine.

The ranks above the quintillion are processed as a double value, but
below this rank the method simply delegates to the long-spelling one
(once more).