Fix some bugs in doubles rank loop
The boundaries of the loop index are narrowed by one on each side: the
high boundary is decreased because it was too high (it did not account
for the decrement in the rank name method), and the low boundary is
increased (by changing the greater-or-equal comparison to a strict
greater one) because it was redundant with the boundary loop in the
long processing method (to which it now falls through for this specific
rank instead).
The comparison between the rank and the value is no longer a strict
comparison, because we want to enter the block also if the value is
exactly equal to the rank (in which case the result will be simply "one
XXXillion").