Reformat code to use a words list
The code structure is now theoretically correct. However, due to IEEE754
doubles not being exact in higher ranks (10^306 is not exactly
representable, for example), the modulo operator does not produce the
expected result.
The code thus needs an extensive rework. This is the first step of it;
functionally, no improvement is visible but it introduces a way of
storing the transcription that will be crucial for further steps: the
words of the transcriptions of the number's integral part are stored in
a list, which is then iterated and appended to the original string
builder.
The list will be later used to store the transcription *in reverse*, and
the ranks in the number also processed in reverse (from lowest to
highest) with successive divisions by one thousand to avoid losing exact
representation (as doubles are precise for low values).