Invert order of number processing
The number value is nibbled from the low end, thousand group by thousand
group, instead of from the high end, because IEEE-754 doubles lose
precision in high values and less so in low ones.
The algorithm now almost works--it is fine for longs, but fails for high
doubles. It is due due to inexact rounding in the division by 1000. I'll
see what I can do to fix this...