Fix the output of some high thousand groups
For example, the millions group of 2^63 was spelled out as 776 instead
of 775, due to the doubles not being able to represent odd integers at
this scale. Now the thousand group is clipped on both sides (divided by
1000, and modulo 1000000) before being spelled out.
This does not fix the output for even bigger numbers, but it's still an
improvement.