Update high test values to IEEE-754 exact numbers
The test numbers in the list above one sextillion and below (the
numerical representation of) Double.MAX_VALUE are not exactly
representable as IEEE-754 doubles. They are updated to the first number
number greater than them that is so.
This has the additional of introducing extra test cases for lower
thousand ranks.
We can note that these values are approximately equal to the original
one to fifteen decimal places: such precision is quite impressive of the
IEE-754 standard.
Also, some Conway-Wechsler systematic rank names are updated in the
transcription of Double.MAX_VALUE.