Use big decimal numbers instead of doubles
When out of the long range, the number is wrapped in a BigDecimal object
before being manipulated, as this class ensures exactitude in the
representation of the result of operations such as division and
remainder.

The changes are minimal: all double operations are changed to BigDecimal
method calls; no refactoring is done.