Add test numbers list w/ transcriptions in French
This is the first step towards providing full support for French--mostly
France-French, but other French regional variations would behave quite
similarly, and thus would require minimal adaptations.

I surmise that the implementation for French will be more complex than
the English one (which was already its own can of worms), not only
because of the traditional use of the long scale (alternation of the
suffixes -ion and -iard for multiples of one thousand), but also because
of the necessary handling of plural forms. Numerals up to *mille* are
generally invariable, except for *vingt* and *cent*, which take the
plural when they are not followed by another numeral adjective: one
writes *quatre-vingt* but *quatre-vingt-un*, *cent* and *deux cents* but
*deux cent un*. On the other hand, the long scale rank names are nouns,
and so require pluralization when appropriate: *un milliard*, and then
*deux milliards*. Also because they are nouns, *vingt* and *cent* can
still take the plural when placed before them: *quatre-vingts millions*.
And what about those pesky dashes all over the place?

This is gonna be fun.