Externalize project properties from subproject POM
The group and artifact IDs, project and Maven compiler release versions
are no longer statically hard-coded in the benchmarks subproject POM,
but replaced with @-delimited placeholders. The actual values will be
injected from the parent POM by the invoker plugin.
This way, the subproject is totally independent from the parent project.