Add parser test w/ list of invalid transcriptions
Like for the valid transcriptions, the list is externalized into a
resource file (named as .txt and not .csv as it has one field per line).

This is one of those rare, blessed occasions where imagination truly is
the only limit.

Disclaimer: the transcriptions listed are only "invalid" in that they
are not accepted by this parser's implementation; this is not a
judgement on their linguistical or historical value: for examples, the
"milliard" rank from the long scale, Conway-Guy-Wechsler's divisive
"quinquadecillion" or the (established, but improper for systematic
usage) "sexdecillion" and "novemdecillion" rank names.