Replace HashMap instances w/ unnamed Map.of calls
The numbers-and-transcriptions maps are instantiated not through a
HashMap constructor, but with a Map.of call with the full list of keys
and values as parameters. This call creates an unmodifiable map
instance, which is be more adapted to this case, as the content of the
maps will not change through the lifetime of the objects.