Begin parsing multiple-word transcriptions
This will be the crux of the implementation.
A handful of variables are added to the method, some regarding the
splitting of the full string into distinct words (substring indices),
some regarding the parsing itself: one double serves as the global
accumulator that is returned at the end; another is used as a temporary
value holder, because the next word might (and, with high probability,
will) be a modifier of this value (actually, a power of ten multiplier;
not all of these multipliers behave the same as there can be up to two
of them consecutive, the first example being "one hundred thousand").