Begin handling the decimal part of the number
A if block is added before the end of the parse method, checking the
loop index. If it is not the value we would expect had it read the whole
array, we consider that it had stopped due to have encountered the
decimal separator.
For now, only the special case of a trailing separator (a separator
specified, but no digit after it) is handled.