Speed up benchmarks subproject build
The build is made faster by skipping the download from the Maven central
repo. Instead, dependencies are fetched from the user's local
repository, so the "downloads" are really file copy operations.
This is made possible by declaring the local repository in a Maven
settings file that is added to the project and referenced in the
configuration of the invoker plugin.
As the settings file is interpolated at build time, its interpolated
copy is marked as ignored.