Change the way the benchmarks are run
Running the benchmarks from a post-build script, although working, felt
wrong--as if it was just an afterthought of the build, while in fact it
is the most important part of it.
The hook script is removed, and instead the benchmarks are run using the
MojoHaus exec plugin, during phase verify. It now feels more like an
integral part of the process.
A properties file is added for the invoker; it currently only contain
one property, that specifies to run the compilation process until the
verify phase (by default, it only runs to package).