Abstract away the words array in an iterator
The words array is replaced with an Iterator<String>. Apart from the
changes from array indexing to Iterator method calls, this has a few
impacts on the structure of the code: the trailing decimal separator
check is moved in the while loop to avoid recomparing the current word,
and the internal method parsing the decimal part now only accepts as
parameter the iterator itself. It also has a new local variable to count
the number of decimal digits, to replace the removed parameter.

This is the first step towards iterating the words *without* a
String.split call, but using an ad-hoc Iterator implementation.