Skip benchmarks with a single build property
In the same vein as skipTests to not run unit tests, the property
skipBenchmarks, when  defined, makes the build process not run the
benchmarks.
It is done by moving the invoker plugin configuration inside the
declaration of a new build profile, the activation of which is commanded
by the absence of this property (double negation. A bit confusing, but I
couldn't find a simpler way).