Prepare parsing of big *illion ranks using regex
When the thousands rank name cannot be found in the list of predefined
ones, the code delegates to a new method parsing it as a bigger (as in,
out of the list range) rank name, instead of raising an error
immediately. This method is declared public to be accessible to test
classes, but it will be hidden away (or even inlined in the calling
method) when the implementation is done.
The method will use regex matching to parse the rank name; a constant
Pattern field is declared, but its expression not yet given (and it
promises to be a huge one). A Matcher object is already defined in the
new method, but not really acted upon, as this will depend upon the
expression itself (although I can already figure that there will be
three capturing groups to be retrieved and converted to digits).