Add test list with invalid big rank names
... and the corresponding test method.

This will ensure that the implementation is strict in what it processes.
Some names currently considered "invalid" might be allowed later on:
sexdecillion, novemdecillion whose usage is somewhat spread (and are
present in dictionaries) although they do not conform to the CGW naming
system, and all quinqua*illions, that are defined by CGW but do not
match proper latin numbers.