Compute the rank number from extracted groups
A map of unit digit names and a list of ten digit names are added (the
first is a map because there are still some digits missing).
These structures are used to convert a rank name component into the
corresponding number, using the Map.get and List.indexOf methods
respectively; these method calls are not checked because due to the
regex match being checked, we are sure that the groups are one of the
listed values.