Handle all digits in rank names
Three, six, seven and nine are now listed in the regex pattern. The
interstitial letters are simply listed, without conformity check on the
next component of the name; due to this, a lot of invalid rank name
tests fail. On the upside, all valid rank names tests succeed, up to 99.