Do not capture the interstitial letters
The group capturing the unit rank component is now extremely simplified:
the processing of the interstitial letter is extracted from it to its
own sequence in the pattern.
This allows to remove the duplicate digits from the units map.
It also  makes the pattern structurally more intuitive: extract the
units digit, then process (and discard) an optional interstitial letter,
the extract the tens digit.
The processing of the interstitial letter is also rather straightforward
in its structure, although the resulting expression is really complex.
It is exploded and indented to improve clarity as much as doable...