Compress lookahead sequences...
... in interstitial letters processing. The tens names are no longer
fully written, but only what is necessary to distinguish them (notably,
the gint suffix is removed). The prefixes for six and seven however are
merged, since the two are treated the same.
The prefixes used are also common with hundreds names, so these lines
will not need to be reworked when this is implemented.