Increase test range to the 110th rank
Valid rank names are added to the test file, which is renamed because it
no longer represented the list.
"Invalid" rank names are also added: some of them are in fact perfectly
valid, but currently out of the range currently handled by the
implementation, and so are considered invalid for the time being.