Use a BigDecimal for the parsing accumulator
This fixes the issue of some big number transcriptions failing to be
parsed, but is clearly lacking in terms of optimization:
BigDecimal.valueOf(double) internally converts the double to a String,
which it then parses; this is unacceptably expensive, especially in a
loop. It will have to be reworked, probably using BigInteger instead,
because the values parsed in this location are integral numbers, and
because the latter's valueOf does not use a String conversion.