Get rid of BigDecimal entirely
The addition of the decimal part (if any) is now done with a simple
primitive addition, without using an intermediat BigDecimal. It turns
out that this does not fail any test; the number range of the tests may
not be thorough, but I am quite certain that this change will not fail
on *more* values that before--should they fail, they would anyway, with
the decimal part being a primitive double.