Use an integer intermediate accumulator
This avoids casts to long when converting the value to a BigInteger
slight readability increase), and is a better-suited type anyway as the
value is necessary integral, and in the thousands at the maximum (so
well within the int range).
The processWord method is also retyped as returning an int value.